comm-aware-sc/requirements.txt
kubernetes==26.1.0
prometheus_api_client==0.5.3
PyYAML==6.0
# kiali==0.3.3